[API] refacto extrract parser in its own file
